Update criteria for regenerating changed methods
Adding additional logic to work out when to re-generate the builder methods. If the
contents of the existing method match any of the possible builder bodies, then the
method will be re-generated, otherwise the method is preserved. This makes it possible
to re-generate a method with a parameter that was previously marked as null to not null.